Add abstract/bio for Brainhack Warsaw & ICHBD
Added abstract for talks at Brainhack Warsaw and for the international
conference on human brain development in Nanning, China.

Also added a passport photo for my visa application to China and
renamed my opencon picture to headshot for ease of emailing.

Finally, I fixed a few typos (well - one typo that had propagated
through loads of copy and pastes of the same bio) and re-arranged the
table of contents to highlight the month of the event.

Fun times :raised_hands: